:rotating_light: move calcite-date-picker to calcite-date
