+++ src/components/calcite-date/calcite-date.scss
