+++ src/demos/calcite-date.html
@@ -28,35 +28,35 @@
   <h1>Calcite Date</h1>
 
   <h2>Standard</h2>
-  <calcite-date-picker></calcite-date-picker>
+  <calcite-date></calcite-date>
 
   <h2>Initial Value (from ISO 8601 string)</h2>
-  <calcite-date-picker value="2020-03-07"></calcite-date-picker>
+  <calcite-date value="2020-03-07"></calcite-date>
 
   <h2>Min &amp; Max</h2>
-  <calcite-date-picker min="2020-03-07" max="2020-03-15" value="2020-04-01"></calcite-date-picker>
+  <calcite-date min="2020-03-07" max="2020-03-15" value="2020-04-01"></calcite-date>
 
   <h2>Start day of locale</h2>
-  <calcite-date-picker locale="es"></calcite-date-picker>
+  <calcite-date locale="es"></calcite-date>
 
   <h2>Locale formatted dates</h2>
-  <calcite-date-picker value="2019-08-23" min="2019-08-09" max="2021-12-18" locale="th"></calcite-date-picker>
+  <calcite-date value="2019-08-23" min="2019-08-09" max="2021-12-18" locale="th"></calcite-date>
 
   <div dir="rtl">
     <h2>Right to left, arabic numerals</h2>
-    <calcite-date-picker value="2000-11-27" locale="ar-EG" dir="rtl"></calcite-date-picker>
+    <calcite-date value="2000-11-27" locale="ar-EG" dir="rtl"></calcite-date>
   </div>
 
   <h2>Always open</h2>
-  <calcite-date-picker value="2000-11-27" no-calendar-input="true" show-calendar="true"></calcite-date-picker>
+  <calcite-date value="2000-11-27" no-calendar-input="true" show-calendar="true"></calcite-date>
 
   <h2>Input Proxy</h2>
-  <calcite-date-picker>
+  <calcite-date>
     <input type="date" value="2000-11-27" min="1999-12-02" max="2020-12-02">
-  </calcite-date-picker>
+  </calcite-date>
 
   <h2>Events</h2>
-  <calcite-date-picker id="dateSample"></calcite-date-picker>
+  <calcite-date id="dateSample"></calcite-date>
   <p>Date selected: <pre><code id="dateDisplay">n/a</code></pre></p>
   <script>
     var dateSample = document.getElementById("dateSample");
@@ -68,10 +68,10 @@
 
   <div class="demo-background-dark">
     <h1>Dark theme</h1>
-    <calcite-date-picker theme="dark" value="2019-08-23"></calcite-date-picker>
+    <calcite-date theme="dark" value="2019-08-23"></calcite-date>
 
     <h2>Always open</h2>
-    <calcite-date-picker value="2000-11-27" theme="dark" no-calendar-input="true" show-calendar="true"></calcite-date-picker>
+    <calcite-date value="2000-11-27" theme="dark" no-calendar-input="true" show-calendar="true"></calcite-date>
   </div>
 </body>
