+++ stencil.config.ts
@@ -10,7 +10,7 @@ export const config: Config = {
     { components: ["calcite-card"] },
     {
       components: [
-        "calcite-date-picker",
+        "calcite-date",
         "calcite-date-month",
         "calcite-date-month-header",
         "calcite-date-day"
